+++ src/filetransfer/base.py
@@ -1,8 +1,9 @@
 """Base classes for source and target implementations."""
 
-import fnmatch
 import logging
+import os
 from contextlib import suppress
+from fnmatch import fnmatch
 
 _logger = logging.getLogger(__name__)
 
@@ -15,10 +16,8 @@ class Endpoint:
     :param str path: the path
     """
 
-    def __init__(self, path, collect_data):
+    def __init__(self, path):
         self._path = path.rstrip('/')
-        self.error_cnt = 0
-        self.error_files = [] if collect_data else None
 
     def __enter__(self):
         return self
@@ -28,13 +27,6 @@ class Endpoint:
 
     def _close(self):
         raise NotImplementedError(f'{self.__class__.__name__}._close()')
-
-    def _handle_error(self, text, file_path, ex):
-        file = file_path[len(self._path) + 1:]
-        self.error_cnt += 1
-        if self.error_files is not None:
-            self.error_files.append((file, str(ex).split('\n')[0]))
-        _logger.error('%s: %s (%s)', text, file, ex)
 
 
 class BaseSource(Endpoint):
@@ -54,8 +46,7 @@ class BaseSource(Endpoint):
     """
 
     def __init__(self, job_cfg):
-        super().__init__(job_cfg['source', 'path'],
-                         job_cfg['job', 'collect_data'])
+        super().__init__(job_cfg['source', 'path'])
         _logger.info('Source: %s', job_cfg['source_url'])
         self._files = job_cfg['source', 'files']
         self._ignore = job_cfg['source', 'ignore']
@@ -74,7 +65,7 @@ class BaseSource(Endpoint):
 
     def _match(self, patterns, path):
         for p in patterns:
-            if fnmatch.fnmatch(path, p):
+            if fnmatch(path, p) or fnmatch(os.path.join(path, ''), p):
                 return True
         return False
 
@@ -83,12 +74,11 @@ class BaseSource(Endpoint):
             p = self._path_join(path, entry)
             if self._recursive and self._isdir(p):
                 try:
-                    for file_path in self._walk(p):
-                        yield file_path
+                    yield from self._walk(p)
                 except Exception as ex:
-                    self._handle_error('Source - directory', p, ex)
+                    yield p, ex
             elif self._isfile(p):
-                yield p
+                yield p, None
 
     def files(self):
         """Return an iterator that yields 2-tuples.
@@ -102,17 +92,22 @@ class BaseSource(Endpoint):
         files = self._patterns(self._files)
         ignore = self._patterns(self._ignore)
         path_len = len(self._path) + 1
-        for file_path in self._walk(self._path):
+        for file_path, exc in self._walk(self._path):
             if (self._match(files, file_path) and
                     not self._match(ignore, file_path)):
                 _logger.debug('source files file_path=%s', file_path)
+                if exc is not None:
+                    _logger.error('Source: %s (%s)', file_path, exc)
+                    yield file_path[path_len:], exc
+                    continue
                 try:
                     yield file_path[path_len:], self._open(file_path, 'rb')
                     if self._delete:
                         with suppress(Exception):
                             self._remove(file_path)
                 except Exception as ex:
-                    self._handle_error('Source - file', file_path, ex)
+                    _logger.error('Source: %s (%s)', file_path, ex)
+                    yield file_path[path_len:], ex
 
 
 class BaseTarget(Endpoint):
@@ -134,8 +129,7 @@ class BaseTarget(Endpoint):
     """
 
     def __init__(self, job_cfg):
-        super().__init__(job_cfg['target', 'path'],
-                         job_cfg['job', 'collect_data'])
+        super().__init__(job_cfg['target', 'path'])
         _logger.info('Target: %s', job_cfg['target_url'])
         self._temp = job_cfg['target', 'temp']
 
@@ -152,29 +146,28 @@ class BaseTarget(Endpoint):
                 return self._path_join(d, self._path_base(file_path))
         return None
 
-    def store(self, file_path, reader):
+    def store(self, path, reader):
         """Save file.
 
         See: :meth:`BaseSource.files`.
 
-        :param str file_path: file path relative to the target base path
+        :param str path: file path relative to the target base path
         :param reader: file reader
         :type reader: :term:`binary file` opened in read-mode
         """
         try:
-            full_path = self._path_join(self._path, file_path)
-            _logger.debug('target store full_path=%s', full_path)
-            self._makedirs(self._path_dir(full_path))
-            tmp_path = self._temp_path(full_path)
+            file_path = self._path_join(self._path, path)
+            _logger.debug('target store file_path=%s', file_path)
+            self._makedirs(self._path_dir(file_path))
+            tmp_path = self._temp_path(file_path)
             _logger.debug('target store tmp_path=%s', tmp_path)
-            if self._path_exists(full_path):
-                self._remove(full_path)
-            with self._open(tmp_path or full_path, 'wb') as fh:
+            if self._path_exists(file_path):
+                self._remove(file_path)
+            with self._open(tmp_path or file_path, 'wb') as fh:
                 while fh.write(reader.read(_CHUNK_SIZE)):
                     pass
             if tmp_path:
-                self._rename(tmp_path, full_path)
-            return True
+                self._rename(tmp_path, file_path)
         except Exception as ex:
-            self._handle_error('Target - file', full_path, ex)
-        return False
+            _logger.error('Target: %s (%s)', file_path, ex)
+            return ex