+++ src/filetransfer/exceptions.py
@@ -1,43 +1,63 @@
 """Exceptions module."""
 
+from .const import ExitCodes
+
 
 class Error(Exception):
     """Base error."""
 
 
 class ConfigError(Error):
-    """Raised when there is a problem with the configuration."""
+    """Raised when there is a problem with the configuration.
 
-    code = 1  #: Status code
+    .. versionchanged:: 0.10.0 ``code`` 1 -> 3
+    """
+
+    code = ExitCodes.CONFIG.code  #: Status code
 
 
 class ConnectError(Error):
     """Raised when there is a problem connecting to a server."""
 
-    code = 2  #: Status code
+    code = ExitCodes.FAILURE.code  #: Status code
 
 
 class TransferError(Error):
-    """Raised when there is a fatal problem during transfer."""
+    """Raised when there is a fatal problem during transfer.
 
-    code = 3  #: Status code
+    .. versionchanged:: 0.10.0 ``code`` 3 -> 2
+    """
+
+    code = ExitCodes.FAILURE.code  #: Status code
 
 
 class SingleInstanceError(Error):
     """Raised when a single instance requirement is violated.
 
     .. versionadded:: 0.9.0
+    .. versionchanged:: 0.10.0 ``code`` 4 -> 0
     """
 
-    code = 4  #: Status code
+    code = ExitCodes.SUCCESS.code  #: Status code
+
+
+class NotReadyError(Error):
+    """Raised when a ready file is required but not present.
+
+    .. versionadded:: 0.10.0
+    """
+
+    code = ExitCodes.SUCCESS.code  #: Status code
 
 
 class Terminated(BaseException):
-    """Raised when terminated with SIGTERM, SIGINT or KeyboardInterrupt.
+    """Raised when terminated with SIGTERM.
 
     This is a subclass of :exc:`BaseException`.
 
-    See :ref:`example <ref-configure-and-run>`.
+    See: :func:`filetransfer.set_sigterm_handler`
+
+    .. versionchanged:: 0.10.0 ``code`` 8 -> 5
     """
 
-    code = 8  #: Status code
+    code = ExitCodes.TERMINATED.code  #: Status code