+++ src/filetransfer/job.py
@@ -1,39 +1,41 @@
 """Job module."""
 
 import logging
-from collections import namedtuple
-from contextlib import nullcontext
-from datetime import datetime
+import os
+import time
+from contextlib import nullcontext, suppress
+from dataclasses import dataclass, field
+from datetime import datetime, timedelta
 
 from salmagundi.utils import ensure_single_instance, AlreadyRunning
 
-from .const import ErrorsEnum
+from .const import ExitCodes, FileTags
 from .exceptions import (ConnectError, TransferError, SingleInstanceError,
-                         Terminated)
+                         NotReadyError, Terminated, Error)
 from .local import LocalSource, LocalTarget
 from .ftp import FTPSource, FTPTarget
 from .sftp import SFTPSource, SFTPTarget
 
+_RETRY_MAX_INTERVAL = 60.0  # seconds
+_RETRY_BACKOFF_FACTOR = 1.0
+_RETRY_BACKOFF_BASE = 2.0
+
 _logger = logging.getLogger(__name__)
 
-JobResult = namedtuple('JobResult', 'files_cnt, src_error_cnt, '
-                       'tgt_error_cnt file_list')
-JobResult.__doc__ = """Class that contains job results.
 
-This class has the following fields:
+@dataclass(frozen=True)
+class JobResult:
+    """JobResult class; documented in api.rst."""
 
-=================  ===
-**files_cnt**      number of successfully transferred files
-**src_error_cnt**  number of files that could not be read
-**tgt_error_cnt**  number of files that could not be written
-**file_list**      list of tuples: (path, info, tag)
-                    - path: path relative to source/target directory
-                    - info: duration of transfer or error text
-                    - tag: > (source), < (target), = (transferred)
+    files_cnt: int
+    src_error_cnt: int
+    tgt_error_cnt: int
+    file_list: list = field(repr=False)
 
-                   if data collection is disabled this will be ``None``
-=================  ===
-"""
+    def __str__(self):
+        return (f'{self.files_cnt} file(s) transferred, '
+                f'{self.src_error_cnt} source error(s), '
+                f'{self.tgt_error_cnt} target error(s)')
 
 
 def run(app_cfg, job_cfg, exc=None):
@@ -44,68 +46,85 @@ def run(app_cfg, job_cfg, exc=None):
     :param job_cfg: the job configuration
     :type job_cfg: salmagundi.config.Config
     :param Exception exc: Exception to be reraised
+    :returns: job result and exit code (0: success, 1: with errors)
+    :rtype: JobResult, bool
     :raises filetransfer.ConnectError: if there is a connection problem
     :raises filetransfer.TransferError: if there is a fatal problem
                                         during transfer
     :raises filetransfer.SingleInstanceError: if a single instance requirement
                                               is violated
     :raises filetransfer.Terminated: if terminatated
-    :raises Exception: if another error occurs
+    :raises Exception: if parameter ``exc`` is set
     """
-    job_id = job_cfg['job_id']
-    if job_cfg['job', 'single_instance']:
-        ctx = ensure_single_instance(job_id,
-                                     lockdir=app_cfg['global', 'locks_dir'],
-                                     err_code=None,
-                                     err_msg=f'already running: job {job_id}')
-    else:
-        ctx = nullcontext()
+    ready_file = _check_ready_file(job_cfg)
+    ctx = _create_context(app_cfg, job_cfg)
+    collect_data = job_cfg['job', 'collect_data']
+    files, exit_code, result = {}, None, None
     try:
         with ctx:
             app_cfg['log_handler'].activate()
-            _run(app_cfg, job_cfg, exc)
+            if exc and isinstance(exc, BaseException):
+                raise exc
+            for i in range(1 + job_cfg['job', 'retries']):
+                try:
+                    transfer(job_cfg, files)
+                    result = create_result(files, collect_data)
+                    _logger.info('Transfer completed: %s', result)
+                    if result.src_error_cnt or result.tgt_error_cnt:
+                        exit_code = ExitCodes.ERRORS
+                    else:
+                        exit_code = ExitCodes.SUCCESS
+                        _remove_ready_file(ready_file)
+                    return result, exit_code.code
+                except Error as ex:
+                    if i == job_cfg['job', 'retries']:
+                        raise
+                    t = min(_RETRY_MAX_INTERVAL,
+                            _RETRY_BACKOFF_FACTOR * _RETRY_BACKOFF_BASE ** i)
+                    _logger.error('Error: %s; retry in %.1f seconds', ex, t)
+                    time.sleep(t)
     except AlreadyRunning as ex:
         raise SingleInstanceError(ex) from None
-
-
-def _run(app_cfg, job_cfg, exc):
-    try:
-        if exc and isinstance(exc, BaseException):
-            raise exc
-        result = transfer(job_cfg)
-        _logger.info('Transfer completed: %d files transferred, %d source '
-                     'errors, %d target errors', *result[:3])
-        if result.src_error_cnt or result.tgt_error_cnt:
-            err = ErrorsEnum.FILES
-        else:
-            err = ErrorsEnum.NONE
-    except ConnectError as ex:
-        _logger.critical('Connect error: %s', ex)
-        err = ErrorsEnum.CONNECT
-        result = ex
-        raise ex
-    except TransferError as ex:
-        _logger.critical('Transfer error: %s', ex)
-        err = ErrorsEnum.TRANSFER
-        result = ex
-        raise ex
-    except (KeyboardInterrupt, Terminated) as ex:
-        _logger.critical('Terminated')
-        err = ErrorsEnum.OTHER
-        result = ex
-        raise Terminated
+    except KeyboardInterrupt as ex:
+        _log_critical('KeyboardInterrupt', ex)
+        exit_code = ExitCodes.TERMINATED
+        result = _add_result(Terminated('KeyboardInterrupt'),
+                             files, collect_data)
+        raise result
+    except Terminated as ex:
+        _log_critical('Terminated', ex)
+        exit_code = ExitCodes.TERMINATED
+        result = _add_result(ex, files, collect_data)
+        raise result
+    except (ConnectError, TransferError) as ex:
+        _log_critical(ex.__class__.__name__, ex)
+        exit_code = ExitCodes.FAILURE
+        result = _add_result(ex, files, collect_data)
+        raise result
     except Exception as ex:
-        _logger.critical('Another error\n%s', exc_info=True)
-        err = ErrorsEnum.OTHER
-        result = ex
-        raise ex
+        _log_critical('Error', ex)
+        result = _add_result(ex, files, collect_data)
+        raise result
     finally:
-        end_time = datetime.now()
-        if app_cfg['mail_config_ok']:
-            from . import mail
-            mail.send(app_cfg, job_cfg, end_time, err, result)
-        _logger.info('Job "%s" finished: duration=%s',
-                     job_cfg['job_id'], end_time - app_cfg['start_time'])
+        if result is not None:
+            end_time = datetime.now()
+            if app_cfg['mail_config_ok']:
+                from . import mail
+                mail.send(app_cfg, job_cfg, end_time, exit_code, result)
+            _logger.info('Job "%s" finished: duration=%s; exit_code=%s',
+                         job_cfg['job_id'],
+                         end_time - app_cfg['start_time'],
+                         exit_code.code if exit_code else None)
+
+
+def _add_result(ex, files, collect_data):
+    ex.result = create_result(files, collect_data)
+    return ex
+
+
+def _log_critical(msg, exc):
+    _logger.critical('%s: %s', msg, exc)
+    _logger.debug(msg, exc_info=exc)
 
 
 def _create_source(job_cfg):
@@ -136,43 +155,102 @@ def _create_target(job_cfg):
         return LocalTarget(job_cfg)
 
 
-def transfer(job_cfg):
+def transfer(job_cfg, files):
     """Transfer files.
+
+    ``files``: path -> duration|(True|False, exc) -- True: src, False: tgt
 
     :param job_cfg: the job configuration
     :type job_cfg: salmagundi.config.Config
+    :param dict files: files
     :return: job result
     :rtype: JobResult
     :raises filetransfer.ConnectError: if there is a connection problem
     :raises filetransfer.TransferError: if there is a fatal problem
                                         during transfer
-    :raises Exception: if another error occurs
     """
-    transferred_files = [] if job_cfg['job', 'collect_data'] else None
-    with _create_source(job_cfg) as source, _create_target(job_cfg) as target:
-        files_cnt = 0
+    with _create_source(job_cfg) as src, _create_target(job_cfg) as tgt:
         try:
-            for file_path, reader in source.files():
+            for file_path, obj in src.files():
                 try:
+                    if (isinstance(files.get(file_path), timedelta) or
+                            file_path == job_cfg['job', 'ready_file']):
+                        continue
+                    if isinstance(obj, Exception):
+                        files[file_path] = (True, obj)
+                        continue
                     start_time = datetime.now()
-                    if target.store(file_path, reader):
-                        files_cnt += 1
+                    exc = tgt.store(file_path, obj)
+                    if exc is None:
                         duration = datetime.now() - start_time
-                        if transferred_files is not None:
-                            transferred_files.append((file_path, duration))
+                        files[file_path] = duration
                         _logger.info('Transferred - file: %s (%s)',
                                      file_path, duration)
+                    else:
+                        files[file_path] = (False, exc)
                 finally:
-                    reader.close()
+                    with suppress(Exception):
+                        obj.close()
         except Exception as ex:
             raise TransferError(ex)
-    if job_cfg['job', 'collect_data']:
+
+
+def create_result(files, collect_data):
+    """Create job result."""
+    transf_cnt, srcerr_cnt, tgterr_cnt = 0, 0, 0
+    if collect_data:
         file_lst = []
-        for lst, tag in zip((transferred_files, source.error_files,
-                             target.error_files), ('=', '>', '<')):
-            for path, info in lst:
-                file_lst.append((path, info, tag))
-        file_lst.sort()
     else:
         file_lst = None
-    return JobResult(files_cnt, source.error_cnt, target.error_cnt, file_lst)
+    for path, value in files.items():
+        if isinstance(value, tuple):
+            if value[0]:
+                srcerr_cnt += 1
+                tag = FileTags.SRCERR
+            else:
+                tgterr_cnt += 1
+                tag = FileTags.TGTERR
+            info = str(value[1]).split('\n')[0]
+        else:
+            transf_cnt += 1
+            info = value
+            tag = FileTags.TRANSF
+        if file_lst is not None:
+            file_lst.append((path, info, tag))
+    if file_lst:
+        file_lst.sort()
+    return JobResult(transf_cnt, srcerr_cnt, tgterr_cnt, file_lst)
+
+
+def _check_ready_file(job_cfg):
+    if job_cfg['job', 'ready_file']:
+        ready_file = os.path.join(job_cfg['source', 'path'],
+                                  job_cfg['job', 'ready_file'])
+        if not os.path.exists(ready_file):
+            raise NotReadyError(f'ready_file {ready_file!r} does not exist')
+    else:
+        ready_file = None
+    return ready_file
+
+
+def _remove_ready_file(ready_file):
+    if ready_file:
+        with suppress(OSError):
+            os.remove(ready_file)
+            _logger.debug('ready_file removed: %s', ready_file)
+
+
+def _create_context(app_cfg, job_cfg):
+    if job_cfg['job', 'single_instance']:
+        if job_cfg['job', 'single_instance'] is True:
+            name = job_cfg['job_id']
+        else:
+            name = job_cfg['job', 'single_instance']
+        msg = f'already running: job {job_cfg["job_id"]} (lock: {name})'
+        ctx = ensure_single_instance(name,
+                                     lockdir=app_cfg['global', 'locks_dir'],
+                                     err_code=None,
+                                     err_msg=msg)
+    else:
+        ctx = nullcontext()
+    return ctx